requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
idna==2.5
packaging==16.8
pyparsing==2.2.0
PyYAML==3.12
requests==2.11.1
six==1.10.0
tsuruclient==0.3.12
urllib3==1.21.1
